Moved compat mod to codec::framed
