+++ src/codec/mod.rs
@@ -1,11 +1,13 @@
 mod command;
 mod error;
+mod framed;
 mod greeting;
 mod mechanism;
 mod zmq_codec;
 
 pub(crate) use command::{ZmqCommand, ZmqCommandName};
 pub(crate) use error::{CodecError, CodecResult};
+pub(crate) use framed::FramedIo;
 pub(crate) use greeting::ZmqGreeting;
 pub(crate) use zmq_codec::ZmqCodec;
